wcn36xx: Fix power-saving with some APs
Though the wcn36xx firmware sends a standard null data packet with
PS flag when entering power-saving mode, some APs (TPLinks) don't
seem to take care of this announcement. This leads to disconnection
or uncoverable degraded throughput when the AP tries to communicate
with the sta (without knowing it's not listening).

Fix that by sending an additional high priority (7) QoS null function
packet (with PS flag) on power-save enabling.

Signed-off-by: Loic Poulain <loic.poulain@linaro.org>